feat: move CNAME records to DNS stack instead of cert stack
