+++ src/DNSStack.ts
@@ -1,3 +1,4 @@
+import * as crypto from 'crypto';
 import { aws_route53 as Route53, Stack, StackProps, aws_ssm as SSM } from 'aws-cdk-lib';
 import { Construct } from 'constructs';
 import { Configurable } from './Configuration';
@@ -9,7 +10,7 @@ export class DNSStack extends Stack {
   zone: Route53.HostedZone;
   accountRootZone: Route53.IHostedZone;
 
-  constructor(scope: Construct, id: string, _props?: DNSStackProps) {
+  constructor(scope: Construct, id: string, props: DNSStackProps) {
     super(scope, id);
 
     const rootZoneId = SSM.StringParameter.valueForStringParameter(this, Statics.accountRootHostedZoneId);
@@ -25,6 +26,10 @@ export class DNSStack extends Stack {
 
     this.addZoneIdAndNametoParams();
     this.addNSToRootCSPzone();
+
+    if (props.configuration.cnameRecords) {
+      this.addCnameRecords(this.zone, props.configuration.cnameRecords);
+    }
 
   }
 
@@ -57,6 +62,23 @@ export class DNSStack extends Stack {
       zone: this.accountRootZone,
       values: this.zone.hostedZoneNameServers,
       recordName: 'yivi-issue',
+    });
+  }
+
+  /**
+   * Add the CNAME records to the hosted zone that are
+   * provided in the branch specific configuration
+   * @param hostedZone the hosted zone to add the records to
+   * @param cnameRecords configruation property containing the records
+   */
+  addCnameRecords(hostedZone: Route53.IHostedZone, cnameRecords: { [key: string]: string }) {
+    Object.entries(cnameRecords).forEach(entry => {
+      const logicalId = crypto.createHash('md5').update(entry[0]).digest('hex').substring(0, 10);
+      new Route53.CnameRecord(this, `record-${logicalId}`, {
+        zone: hostedZone,
+        recordName: entry[0],
+        domainName: entry[1],
+      });
     });
   }
