+++ test/main.test.ts
@@ -82,7 +82,9 @@ test('StackHasApiGateway', () => {
   const app = new App();
   const keyStack = new KeyStack(app, 'keystack');
   const sessionsStack = new SessionsStack(app, 'test', { key: keyStack.key});
-  new DNSStack(app, 'dns');
+  new DNSStack(app, 'dns', {
+    configuration: config,
+  });
   const stack = new ApiStack(app, 'api', { 
     sessionsTable: sessionsStack.sessionsTable, 
     configuration: config,
@@ -96,7 +98,9 @@ test('StackHasLambdas', () => {
   const app = new App();
   const keyStack = new KeyStack(app, 'keystack');
   const sessionsStack = new SessionsStack(app, 'test', { key: keyStack.key});
-  new DNSStack(app, 'dns');
+  new DNSStack(app, 'dns', {
+    configuration: config,
+  });
   const stack = new ApiStack(app, 'api', { 
     sessionsTable: sessionsStack.sessionsTable, 
     configuration: config,
